Removal of preint.js bundle + proper imports instead of global declarations
