+++ assets/scripts/app/focus.js
@@ -1,6 +1,5 @@
-/* global system */
-
 import { galleryState } from '../gallery/view'
+import { system } from '../preinit/system_capabilities'
 import { fetchStreetForVerification } from '../streets/xhr'
 import { saveSettingsLocally } from '../users/settings'
 import { getAbortEverything } from './initialization'
