+++ assets/scripts/app/page_url.js
@@ -1,6 +1,5 @@
-/* global debug */
-
 import { shareMenu } from '../menus/_share'
+import { debug } from '../preinit/debug_settings'
 import { getStreet, getStreetUrl } from '../streets/data_model'
 import { setMode, MODES } from './mode'
 import {
