+++ assets/scripts/main.js
@@ -3,20 +3,23 @@
  *
  */
 
-import $ from 'jquery'
-window.$ = $
-
-import _ from 'lodash'
-window._ = _
+import Raven from 'raven-js'
 
 // Polyfills
+import 'babel-polyfill'
+import 'whatwg-fetch' // fetch API
+import 'handjs' // microsoft's pointer events / touch-action spec
 import './vendor/canvas-toBlob.js'
 import './vendor/Blob.js'
+import './vendor/modernizr-custom'
+import './vendor/polyfills/customevent' // customEvent in IE
 
 // Main object
 import { Stmx } from './app/initialization'
 window.Stmx = Stmx
 
+import { startListening } from './app/keypress'
+import { system } from './preinit/system_capabilities'
 // import modules for side-effects
 import './app/blocking_shield'
 import './app/debug_info'
@@ -37,6 +40,37 @@ import './streets/name'
 import './streets/scroll'
 import './util/fetch_nonblocking'
 
-import { startListening } from './app/keypress'
+// Error tracking
+// Load this before all other modules. Only load when run in production.
+if (window.location.hostname === 'streetmix.net' || window.location.hostname === 'www.streetmix.net') {
+  Raven.config('https://fac2c23600414d2fb78c128cdbdeaf6f@app.getsentry.com/82756', {
+    whitelistUrls: [/streetmix\.net/, /www\.streetmix\.net/]
+  }).install()
+}
+
+function setScaleForPhone () {
+  var meta = document.createElement('meta')
+  meta.setAttribute('name', 'viewport')
+
+  if (system.phone) {
+    meta.setAttribute('content', 'initial-scale=.5, maximum-scale=.5')
+  } else {
+    meta.setAttribute('content', 'initial-scale=1, maximum-scale=1')
+  }
+
+  var headEls = document.getElementsByTagName('head')
+  headEls[0].appendChild(meta)
+}
+setScaleForPhone()
+
+// This event is fired by _onEverythingLoaded() in the deprecated
+// global bundle. This allows things in the modular bundle to respond
+// to that function without needing to be exported globally.
+// This should eventually not be required & can be removed.
+window.addEventListener('stmx:everything_loaded', function (e) {
+  /* global _onEverythingLoaded2 */
+  _onEverythingLoaded2()
+})
+
 // Start listening for keypresses
 startListening()
