+++ assets/scripts/segments/view.js
@@ -1,9 +1,9 @@
-/* global system, debug */
-
 import { getInitializing } from '../app/initialization'
 import { images, getImagesToBeLoaded } from '../app/load_resources'
 import { msg } from '../app/messages'
 import { infoBubble, INFO_BUBBLE_TYPE_SEGMENT } from '../info_bubble/info_bubble'
+import { debug } from '../preinit/debug_settings'
+import { system } from '../preinit/system_capabilities'
 import {
   getStreet,
   saveStreetToServerIfNecessary,
